requirements.txt
PyPDF2
streamlit
python-docx
tempfile
langchain-text-splitters
langchain-google-genai
python-dotenv
nltk
scikit-learn
numpy
faiss-gpu